*: address keyspace observability review comments
