+++ cmd/tidb-server/BUILD.bazel
@@ -68,16 +68,11 @@ go_library(
         "@com_github_opentracing_opentracing_go//:opentracing-go",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_pingcap_kvproto//pkg/keyspacepb",
         "@com_github_pingcap_log//:log",
         "@com_github_prometheus_client_golang//prometheus",
         "@com_github_prometheus_client_golang//prometheus/push",
-        "@com_github_tikv_client_go_v2//config",
         "@com_github_tikv_client_go_v2//tikv",
         "@com_github_tikv_client_go_v2//txnkv/transaction",
-        "@com_github_tikv_pd_client//:client",
-        "@com_github_tikv_pd_client//opt",
-        "@com_github_tikv_pd_client//pkg/caller",
         "@org_uber_go_automaxprocs//maxprocs",
         "@org_uber_go_zap//:zap",
     ],
@@ -122,7 +117,6 @@ go_test(
         "//pkg/sessionctx/vardef",
         "//pkg/sessionctx/variable",
         "//pkg/testkit/testsetup",
-        "@com_github_pingcap_kvproto//pkg/keyspacepb",
         "@com_github_stretchr_testify//require",
         "@io_opencensus_go//stats/view",
         "@org_uber_go_goleak//:goleak",
