+++ cmd/tidb-server/main.go
@@ -19,7 +19,6 @@ import (
 	"flag"
 	"fmt"
 	"io/fs"
-	"maps"
 	"os"
 	"runtime"
 	"strconv"
@@ -31,7 +30,6 @@ import (
 	"github.com/opentracing/opentracing-go"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
-	"github.com/pingcap/kvproto/pkg/keyspacepb"
 	"github.com/pingcap/log"
 	"github.com/pingcap/tidb/pkg/bindinfo"
 	"github.com/pingcap/tidb/pkg/config"
@@ -93,12 +91,8 @@ import (
 	repository "github.com/pingcap/tidb/pkg/util/workloadrepo"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/push"
-	tikvconfig "github.com/tikv/client-go/v2/config"
 	"github.com/tikv/client-go/v2/tikv"
 	"github.com/tikv/client-go/v2/txnkv/transaction"
-	pd "github.com/tikv/pd/client"
-	"github.com/tikv/pd/client/opt"
-	"github.com/tikv/pd/client/pkg/caller"
 	"go.uber.org/automaxprocs/maxprocs"
 	"go.uber.org/zap"
 )
@@ -321,6 +315,7 @@ func main() {
 	}
 
 	var standbyController server.StandbyController
+	var activationMetadata map[string]string
 	if config.GetGlobalConfig().Standby.StandByMode {
 		standbyController = standby.NewLoadKeyspaceController()
 	}
@@ -337,12 +332,15 @@ func main() {
 		defer standbyController.EndStandby(err)
 		// need to validate config again in case of config change via standby
 		terror.MustNil(config.GetGlobalConfig().Valid())
+		if c, ok := standbyController.(*standby.LoadKeyspaceController); ok {
+			activationMetadata = c.ActivationMetadata()
+		}
 	}
 
 	signal.SetupUSR1Handler()
 	err = registerStores()
 	terror.MustNil(err)
-	err = prepareKeyspaceObservability()
+	err = prepareKeyspaceObservability(activationMetadata)
 	terror.MustNil(err)
 	err = metricsutil.RegisterMetrics()
 	terror.MustNil(err)
@@ -1155,78 +1153,38 @@ func closeStmtSummary() {
 	}
 }
 
-var keyspaceMetaComponentName = caller.Component("tidb-keyspace-meta")
-
 const (
 	keyspaceIDMetricLabel   = "keyspace_id"
 	keyspaceNameMetricLabel = "keyspace_name"
 )
 
-func prepareKeyspaceObservability() error {
+func prepareKeyspaceObservability(metadata map[string]string) error {
 	cfg := config.GetGlobalConfig()
-	if !kerneltype.IsNextGen() {
-		return nil
-	}
-	if keyspace.IsKeyspaceNameEmpty(cfg.KeyspaceName) || cfg.Store != config.StoreTypeTiKV {
+	if !kerneltype.IsNextGen() || cfg.Store != config.StoreTypeTiKV {
 		return nil
 	}
 	metricscommon.SetConstLabels(keyspaceNameMetricLabel, cfg.KeyspaceName)
-	pdAddrs, _, _, err := tikvconfig.ParsePath("tikv://" + cfg.Path)
-	if err != nil {
-		return err
-	}
-	timeoutSec := time.Duration(cfg.PDClient.PDServerTimeout) * time.Second
-	pdCli, err := pd.NewClient(keyspaceMetaComponentName, pdAddrs, pd.SecurityOption{
-		CAPath:   cfg.Security.ClusterSSLCA,
-		CertPath: cfg.Security.ClusterSSLCert,
-		KeyPath:  cfg.Security.ClusterSSLKey,
-	}, opt.WithCustomTimeoutOption(timeoutSec), opt.WithInitMetricsOption(false))
-	if err != nil {
-		return err
-	}
-	defer pdCli.Close()
-
-	keyspaceMeta, err := getKeyspaceMeta(pdCli, cfg.KeyspaceName)
-	if err != nil {
-		return err
-	}
-	keyspace.SetKeyspaceMeta(keyspaceMeta)
-	return prepareKeyspaceObservabilityWithKeyspaceMeta(keyspaceMeta, cfg.KeyspaceName, deploymode.IsStarter())
+	return prepareKeyspaceObservabilityWithMetadata(metadata, cfg.KeyspaceName, deploymode.IsStarter())
 }
 
-func getKeyspaceMeta(pdCli pd.Client, keyspaceName string) (*keyspacepb.KeyspaceMeta, error) {
-	var keyspaceMeta *keyspacepb.KeyspaceMeta
-	err := util.RunWithRetry(util.DefaultMaxRetries, util.RetryInterval, func() (bool, error) {
-		var errInner error
-		keyspaceMeta, errInner = pdCli.LoadKeyspace(context.TODO(), keyspaceName)
-		if kvstore.IsNotBootstrappedError(errInner) || kvstore.IsKeyspaceNotExistError(errInner) {
-			return true, errInner
-		}
-		return false, errInner
-	})
-	if err != nil {
-		return nil, err
-	}
-	return keyspaceMeta, nil
-}
-
-func prepareKeyspaceObservabilityWithKeyspaceMeta(keyspaceMeta *keyspacepb.KeyspaceMeta, keyspaceName string, includeConfiguredFields bool) error {
-	if keyspaceMeta == nil {
-		return nil
-	}
+func prepareKeyspaceObservabilityWithMetadata(metadata map[string]string, keyspaceName string, includeConfiguredFields bool) error {
 	resolvedValues := config.KeyspaceObservabilityValues{
 		MetricLabels: map[string]string{
-			keyspaceIDMetricLabel:   fmt.Sprint(keyspaceMeta.GetId()),
 			keyspaceNameMetricLabel: keyspaceName,
 		},
 	}
+	if keyspaceID, ok := metadata[keyspaceIDMetricLabel]; ok {
+		resolvedValues.MetricLabels[keyspaceIDMetricLabel] = keyspaceID
+	}
 	if includeConfiguredFields {
 		copiedConfig := *config.GetGlobalConfig()
-		if err := copiedConfig.ResolveKeyspaceObservability(keyspaceMeta.GetConfig()); err != nil {
+		if err := copiedConfig.ResolveKeyspaceObservability(metadata); err != nil {
 			return err
 		}
 		configuredValues := copiedConfig.KeyspaceObservabilityValues.Clone()
-		maps.Copy(resolvedValues.MetricLabels, configuredValues.MetricLabels)
+		for k, v := range configuredValues.MetricLabels {
+			resolvedValues.MetricLabels[k] = v
+		}
 		resolvedValues.SlowLogFields = configuredValues.SlowLogFields
 		resolvedValues.StmtLogFields = configuredValues.StmtLogFields
 	}
