+++ pkg/keyspace/keyspace_test.go
@@ -18,12 +18,9 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/pingcap/kvproto/pkg/keyspacepb"
 	"github.com/pingcap/tidb/pkg/config"
 	"github.com/pingcap/tidb/pkg/config/kerneltype"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/zap"
-	"go.uber.org/zap/zaptest/observer"
 )
 
 func TestSetKeyspaceNameInConf(t *testing.T) {
@@ -72,41 +69,6 @@ func TestNoKeyspaceNameSet(t *testing.T) {
 	} else {
 		require.Nil(t, getKeyspaceNameByte)
 	}
-}
-
-func TestSetKeyspaceMeta(t *testing.T) {
-	restore := config.RestoreFunc()
-	defer restore()
-	t.Cleanup(func() {
-		keyspaceMeta.Delete("test_keyspace")
-	})
-
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.KeyspaceName = ""
-	})
-
-	meta := &keyspacepb.KeyspaceMeta{Id: 42, Name: "test_keyspace"}
-	SetKeyspaceMeta(meta)
-
-	got, ok := GetKeyspaceMeta("test_keyspace")
-	require.True(t, ok)
-	require.Equal(t, uint32(42), got.GetId())
-	require.Equal(t, "test_keyspace", got.GetName())
-
-	SetKeyspaceMeta(nil)
-	_, ok = GetKeyspaceMeta("")
-	require.False(t, ok)
-
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.KeyspaceName = "test_keyspace"
-	})
-	core, logs := observer.New(zap.InfoLevel)
-	logger := zap.New(core, WrapZapcoreWithKeyspace())
-	logger.Info("test")
-	entries := logs.All()
-	require.Len(t, entries, 1)
-	require.Equal(t, "test_keyspace", entries[0].ContextMap()["keyspaceName"])
-	require.Equal(t, uint32(42), entries[0].ContextMap()["keyspaceID"])
 }
 
 func BenchmarkGetKeyspaceNameBytesBySettings(b *testing.B) {
